feat: 계좌 송금, 입금 로직 구현
* 계좌 송금,입금, 삭제 추가 ---

+++ src/main/java/com/c9Pay/creditservice/dto/charge/ChargeAmount.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChargeAmount {
-    private Integer creditAmount;
+    private Long creditAmount;
 }
